requirements.txt
python-telegram-bot
requests
pybit
